Refactor request-making out of StripeMethod
Refactor out StripeResource#wrapTimeout as utils.callbackifyPromise
