+++ lib/StripeMethod.basic.js
@@ -2,6 +2,7 @@
 
 var isPlainObject = require('lodash.isplainobject');
 var stripeMethod = require('./StripeMethod');
+var utils = require('./utils');
 
 module.exports = {
 
@@ -52,7 +53,7 @@ module.exports = {
     var urlData = this.createUrlData();
     var path = this.createFullPath('/' + id, urlData);
 
-    return this.wrapTimeout(new Promise((function(resolve, reject) {
+    return utils.callbackifyPromise(new Promise((function(resolve, reject) {
       if (isNull) {
         // Reset metadata:
         sendMetadata(null, auth);
@@ -96,7 +97,7 @@ module.exports = {
     var urlData = this.createUrlData();
     var path = this.createFullPath('/' + id, urlData);
 
-    return this.wrapTimeout(new Promise((function(resolve, reject) {
+    return utils.callbackifyPromise(new Promise((function(resolve, reject) {
       this._request('GET', path, {}, auth, {}, function(err, response) {
         if (err) {
           reject(err);
