+++ lib/StripeMethod.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var utils = require('./utils');
-var OPTIONAL_REGEX = /^optional!/;
+var makeRequest = require('./makeRequest');
 
 /**
  * Create an API method from the declared spec.
@@ -17,106 +17,17 @@ var OPTIONAL_REGEX = /^optional!/;
  *  Usefully for applying transforms to data on a per-method basis.
  */
 function stripeMethod(spec) {
-  var commandPath = typeof spec.path == 'function' ? spec.path
-    : utils.makeURLInterpolator(spec.path || '');
-  var requestMethod = (spec.method || 'GET').toUpperCase();
-  var urlParams = spec.urlParams || [];
-  var encode = spec.encode || function(data) {return data;};
 
   return function() {
     var self = this;
     var args = [].slice.call(arguments);
 
     var callback = typeof args[args.length - 1] == 'function' && args.pop();
-    var urlData = this.createUrlData();
 
-    return this.wrapTimeout(new Promise((function(resolve, reject) {
-      for (var i = 0, l = urlParams.length; i < l; ++i) {
-        var path
-        var err;
+    var requestPromise = makeRequest(self, args, spec, {});
 
-        // Note that we shift the args array after every iteration so this just
-        // grabs the "next" argument for use as a URL parameter.
-        var arg = args[0];
 
-        var param = urlParams[i];
-
-        var isOptional = OPTIONAL_REGEX.test(param);
-        param = param.replace(OPTIONAL_REGEX, '');
-
-        if (param == 'id' && typeof arg !== 'string') {
-          path = this.createResourcePathWithSymbols(spec.path);
-          err = new Error(
-            'Stripe: "id" must be a string, but got: ' + typeof arg +
-            ' (on API request to `' + requestMethod + ' ' + path + '`)'
-          );
-          reject(err);
-          return;
-        }
-
-        if (!arg) {
-          if (isOptional) {
-            urlData[param] = '';
-            continue;
-          }
-
-          path = this.createResourcePathWithSymbols(spec.path);
-          err = new Error(
-            'Stripe: Argument "' + urlParams[i] + '" required, but got: ' + arg +
-            ' (on API request to `' + requestMethod + ' ' + path + '`)'
-          );
-          reject(err);
-          return;
-        }
-
-        urlData[param] = args.shift();
-      }
-
-      var data;
-      try {
-        data = encode(utils.getDataFromArgs(args));
-      } catch (e) {
-        reject(e);
-      }
-      var opts = utils.getOptionsFromArgs(args);
-
-      if (args.length) {
-        path = this.createResourcePathWithSymbols(spec.path);
-        err = new Error(
-          'Stripe: Unknown arguments (' + args + '). Did you mean to pass an options ' +
-          'object? See https://github.com/stripe/stripe-node/wiki/Passing-Options.' +
-          ' (on API request to ' + requestMethod + ' `' + path + '`)'
-        );
-        reject(err);
-        return;
-      }
-
-      var requestPath = this.createFullPath(commandPath, urlData);
-      var options = {headers: Object.assign(opts.headers, spec.headers)};
-
-      if (spec.validator) {
-        try {
-          spec.validator(data, options);
-        } catch (err) {
-          reject(err);
-          return;
-        }
-      }
-
-      function requestCallback(err, response) {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(
-            spec.transformResponseData ?
-              spec.transformResponseData(response) :
-              response
-          );
-        }
-      }
-
-      self._request(requestMethod, requestPath, data, opts.auth, options, requestCallback);
-    }).bind(this)), callback);
+    return utils.callbackifyPromise(requestPromise, callback);
   };
 }
