+++ lib/StripeResource.js
@@ -87,18 +87,8 @@ StripeResource.prototype = {
     return urlData;
   },
 
-  wrapTimeout: function(promise, callback) {
-    if (callback) {
-      // Ensure callback is called outside of promise stack.
-      return promise.then(function(res) {
-        setTimeout(function() { callback(null, res) }, 0);
-      }, function(err) {
-        setTimeout(function() { callback(err, null); }, 0);
-      });
-    }
-
-    return promise;
-  },
+  // DEPRECATED: Here for backcompat in case users relied on this.
+  wrapTimeout: utils.callbackifyPromise,
 
   _timeoutHandler: function(timeout, req, callback) {
     var self = this;
