+++ lib/utils.js
@@ -227,6 +227,19 @@ var utils = module.exports = {
     }
     return false;
   },
+
+  callbackifyPromise: function(promise, callback) {
+    if (callback) {
+      // Ensure callback is called outside of promise stack.
+      return promise.then(function(res) {
+        setTimeout(function() { callback(null, res) }, 0);
+      }, function(err) {
+        setTimeout(function() { callback(err, null); }, 0);
+      });
+    }
+
+    return promise;
+  },
 };
 
 function emitWarning(warning) {
